wasm_encoder

Struct CanonicalFunctionSection

pub struct CanonicalFunctionSection { /* private fields */ }

An encoder for the canonical function section of WebAssembly components.

Example

use wasm_encoder::{Component, CanonicalFunctionSection, CanonicalOption};

let mut functions = CanonicalFunctionSection::new();
functions.lift(0, 0, [CanonicalOption::UTF8]);

let mut component = Component::new();
component.section(&functions);

let bytes = component.finish();

Implementations

impl CanonicalFunctionSection

pub fn new() -> Self

Construct a new component function section encoder.

pub fn len(&self) -> u32

The number of functions in the section.

pub fn is_empty(&self) -> bool

Determines if the section is empty.

pub fn lift<O>( &mut self, core_func_index: u32, type_index: u32, options: O, ) -> &mut Self
where O: IntoIterator<Item = CanonicalOption>, O::IntoIter: ExactSizeIterator,

Define a function that will lift a core WebAssembly function to the canonical ABI.

pub fn lower<O>(&mut self, func_index: u32, options: O) -> &mut Self
where O: IntoIterator<Item = CanonicalOption>, O::IntoIter: ExactSizeIterator,

Define a function that will lower a canonical ABI function to a core WebAssembly function.

pub fn resource_new(&mut self, ty_index: u32) -> &mut Self

Defines a function which will create an owned handle to the resource specified by ty_index.

pub fn resource_drop(&mut self, ty_index: u32) -> &mut Self

Defines a function which will drop the specified type of handle.

pub fn resource_drop_async(&mut self, ty_index: u32) -> &mut Self

Defines a function which will drop the specified type of handle.

pub fn resource_rep(&mut self, ty_index: u32) -> &mut Self

Defines a function which will return the representation of the specified resource type.

pub fn thread_spawn_ref(&mut self, ty_index: u32) -> &mut Self

Defines a function which will spawn a new thread by invoking a shared function of type ty_index.

pub fn thread_spawn_indirect( &mut self, ty_index: u32, table_index: u32, ) -> &mut Self

Defines a function which will spawn a new thread by invoking a shared function indirectly through a funcref table.

pub fn thread_available_parallelism(&mut self) -> &mut Self

Defines a function which will return the number of threads that can be expected to execute concurrently.

pub fn backpressure_inc(&mut self) -> &mut Self

Defines a function which tells the host to increment the backpressure counter.

pub fn backpressure_dec(&mut self) -> &mut Self

Defines a function which tells the host to decrement the backpressure counter.

pub fn task_return<O>( &mut self, ty: Option<ComponentValType>, options: O, ) -> &mut Self
where O: IntoIterator<Item = CanonicalOption>, O::IntoIter: ExactSizeIterator,

Defines a function which returns a result to the caller of a lifted export function. This allows the callee to continue executing after returning a result.

pub fn task_cancel(&mut self) -> &mut Self

Defines a function to acknowledge cancellation of the current task.

pub fn context_get(&mut self, i: u32) -> &mut Self

Defines a new context.get intrinsic of the ith slot.

pub fn context_set(&mut self, i: u32) -> &mut Self

Defines a new context.set intrinsic of the ith slot.

pub fn thread_yield(&mut self, cancellable: bool) -> &mut Self

Defines a function which yields control to the host so that other tasks are able to make progress, if any.

If cancellable is true, the caller instance may be reentered.

pub fn subtask_drop(&mut self) -> &mut Self

Defines a function to drop a specified task which has completed.

pub fn subtask_cancel(&mut self, async_: bool) -> &mut Self

Defines a function to cancel an in-progress task.

pub fn stream_new(&mut self, ty: u32) -> &mut Self

Defines a function to create a new stream handle of the specified type.

pub fn stream_read<O>(&mut self, ty: u32, options: O) -> &mut Self
where O: IntoIterator<Item = CanonicalOption>, O::IntoIter: ExactSizeIterator,

Defines a function to read from a stream of the specified type.

pub fn stream_write<O>(&mut self, ty: u32, options: O) -> &mut Self
where O: IntoIterator<Item = CanonicalOption>, O::IntoIter: ExactSizeIterator,

Defines a function to write to a stream of the specified type.

pub fn stream_cancel_read(&mut self, ty: u32, async_: bool) -> &mut Self

Defines a function to cancel an in-progress read from a stream of the specified type.

pub fn stream_cancel_write(&mut self, ty: u32, async_: bool) -> &mut Self

Defines a function to cancel an in-progress write to a stream of the specified type.

pub fn stream_drop_readable(&mut self, ty: u32) -> &mut Self

Defines a function to drop the readable end of a stream of the specified type.

pub fn stream_drop_writable(&mut self, ty: u32) -> &mut Self

Defines a function to drop the writable end of a stream of the specified type.

pub fn future_new(&mut self, ty: u32) -> &mut Self

Defines a function to create a new future handle of the specified type.

pub fn future_read<O>(&mut self, ty: u32, options: O) -> &mut Self
where O: IntoIterator<Item = CanonicalOption>, O::IntoIter: ExactSizeIterator,

Defines a function to read from a future of the specified type.

pub fn future_write<O>(&mut self, ty: u32, options: O) -> &mut Self
where O: IntoIterator<Item = CanonicalOption>, O::IntoIter: ExactSizeIterator,

Defines a function to write to a future of the specified type.

pub fn future_cancel_read(&mut self, ty: u32, async_: bool) -> &mut Self

Defines a function to cancel an in-progress read from a future of the specified type.

pub fn future_cancel_write(&mut self, ty: u32, async_: bool) -> &mut Self

Defines a function to cancel an in-progress write to a future of the specified type.

pub fn future_drop_readable(&mut self, ty: u32) -> &mut Self

Defines a function to drop the readable end of a future of the specified type.

pub fn future_drop_writable(&mut self, ty: u32) -> &mut Self

Defines a function to drop the writable end of a future of the specified type.

pub fn error_context_new<O>(&mut self, options: O) -> &mut Self
where O: IntoIterator<Item = CanonicalOption>, O::IntoIter: ExactSizeIterator,

Defines a function to create a new error-context with a specified debug message.

pub fn error_context_debug_message<O>(&mut self, options: O) -> &mut Self
where O: IntoIterator<Item = CanonicalOption>, O::IntoIter: ExactSizeIterator,

Defines a function to get the debug message for a specified error-context.

Note that the debug message might not necessarily match what was passed to error-context.new.

pub fn error_context_drop(&mut self) -> &mut Self

Defines a function to drop a specified error-context.

pub fn waitable_set_new(&mut self) -> &mut Self

Declare a new waitable-set.new intrinsic, used to create a waitable-set pseudo-resource.

pub fn waitable_set_wait(&mut self, async_: bool, memory: u32) -> &mut Self

Declare a new waitable-set.wait intrinsic, used to block on a waitable-set.

pub fn waitable_set_poll(&mut self, async_: bool, memory: u32) -> &mut Self

Declare a new waitable-set.wait intrinsic, used to check, without blocking, if anything in a waitable-set is ready.

pub fn waitable_set_drop(&mut self) -> &mut Self

Declare a new waitable-set.drop intrinsic, used to dispose a waitable-set pseudo-resource.

pub fn waitable_join(&mut self) -> &mut Self

Declare a new waitable.join intrinsic, used to add an item to a waitable-set.

pub fn thread_index(&mut self) -> &mut Self

Declare a new thread.index intrinsic, used to get the index of the current thread.

pub fn thread_new_indirect( &mut self, ty_index: u32, table_index: u32, ) -> &mut Self

Declare a new thread.new-indirect intrinsic, used to create a new thread by invoking a function indirectly through a funcref table.

pub fn thread_suspend_to_suspended(&mut self, cancellable: bool) -> &mut Self

Declare a new thread.suspend-to-suspended intrinsic, used to switch execution to another suspended thread.

pub fn thread_suspend(&mut self, cancellable: bool) -> &mut Self

Declare a new thread.suspend intrinsic, used to suspend execution of the current thread.

pub fn thread_suspend_to(&mut self, cancellable: bool) -> &mut Self

Declare a new thread.suspend-to intrinsic, used to suspend the current thread and switch to another thread that might not be suspended.

pub fn thread_unsuspend(&mut self) -> &mut Self

Declare a new thread.unsuspend intrinsic, used to resume execution of the given thread.

pub fn thread_yield_to_suspended(&mut self, cancellable: bool) -> &mut Self

Declare a new thread.yield-to-suspended intrinsic, used to yield execution to a given suspended thread.