Struct wasm_encoder::CodeSection

pub struct CodeSection { /* private fields */ }

An encoder for the code section.

Code sections are only supported for modules.

Example

use wasm_encoder::{
    CodeSection, Function, FunctionSection, Instruction, Module,
    TypeSection, ValType
};

let mut types = TypeSection::new();
types.function(vec![], vec![ValType::I32]);

let mut functions = FunctionSection::new();
let type_index = 0;
functions.function(type_index);

let locals = vec![];
let mut func = Function::new(locals);
func.instruction(&Instruction::I32Const(42));
let mut code = CodeSection::new();
code.function(&func);

let mut module = Module::new();
module
    .section(&types)
    .section(&functions)
    .section(&code);

let wasm_bytes = module.finish();

Implementations

impl CodeSection

pub fn new() -> Self

Create a new code section encoder.

pub fn len(&self) -> u32

The number of functions in the section.

pub fn byte_len(&self) -> usize

The number of bytes already added to this section.

This number doesn’t include the vector length that precedes the code entries, since it has a variable size that isn’t known until all functions are added.

pub fn is_empty(&self) -> bool

Determines if the section is empty.

pub fn function(&mut self, func: &Function) -> &mut Self

Write a function body into this code section.

pub fn raw(&mut self, data: &[u8]) -> &mut Self

Add a raw byte slice into this code section as a function body.

The length prefix of the function body will be automatically prepended, and should not be included in the raw byte slice.

Example

You can use the raw method to copy an already-encoded function body into a new code section encoder:

//                  id, size, # entries, entry
let code_section = [10, 6,    1,         4, 0, 65, 0, 11];

// Parse the code section.
let reader = wasmparser::CodeSectionReader::new(&code_section, 0).unwrap();
let body = reader.into_iter().next().unwrap().unwrap();
let body_range = body.range();

// Add the body to a new code section encoder by copying bytes rather
// than re-parsing and re-encoding it.
let mut encoder = wasm_encoder::CodeSection::new();
encoder.raw(&code_section[body_range.start..body_range.end]);