Struct wasm_encoder::ComponentNameSection

pub struct ComponentNameSection { /* private fields */ }

Encoding for the component-name custom section which assigns human-readable names to items within a component.

Implementations

impl ComponentNameSection

pub fn new() -> Self

Creates a new blank name custom section.

pub fn component(&mut self, name: &str)

Appends a component name subsection to this section.

This will indicate that the name of the entire component should be the name specified. Note that this should be encoded first before other subsections.

pub fn core_funcs(&mut self, names: &NameMap)

Appends a decls name subsection to name core functions within the component.

pub fn core_tables(&mut self, names: &NameMap)

Appends a decls name subsection to name core tables within the component.

pub fn core_memories(&mut self, names: &NameMap)

Appends a decls name subsection to name core memories within the component.

pub fn core_globals(&mut self, names: &NameMap)

Appends a decls name subsection to name core globals within the component.

pub fn core_types(&mut self, names: &NameMap)

Appends a decls name subsection to name core types within the component.

pub fn core_modules(&mut self, names: &NameMap)

Appends a decls name subsection to name core modules within the component.

pub fn core_instances(&mut self, names: &NameMap)

Appends a decls name subsection to name core instances within the component.

pub fn funcs(&mut self, names: &NameMap)

Appends a decls name subsection to name component functions within the component.

pub fn values(&mut self, names: &NameMap)

Appends a decls name subsection to name component values within the component.

pub fn types(&mut self, names: &NameMap)

Appends a decls name subsection to name component type within the component.

pub fn components(&mut self, names: &NameMap)

Appends a decls name subsection to name components within the component.

pub fn instances(&mut self, names: &NameMap)

Appends a decls name subsection to name component instances within the component.

pub fn raw(&mut self, id: u8, data: &[u8])

Appends a raw subsection with the given id and data.

pub fn is_empty(&self) -> bool

Returns whether this section is empty, or nothing has been encoded.

pub fn as_custom<'a>(&'a self) -> CustomSection<'a>

View the encoded section as a CustomSection.