wasm_encoder

Struct ElementSection

pub struct ElementSection { /* private fields */ }

An encoder for the element section.

Element sections are only supported for modules.

Example

use wasm_encoder::{
    Elements, ElementSection, Module, TableSection, TableType,
    RefType, ConstExpr
};

let mut tables = TableSection::new();
tables.table(TableType {
    element_type: RefType::FUNCREF,
    minimum: 128,
    maximum: None,
    table64: false,
    shared: false,
});

let mut elements = ElementSection::new();
let table_index = 0;
let offset = ConstExpr::i32_const(42);
let functions = Elements::Functions(&[
    // Function indices...
]);
elements.active(Some(table_index), &offset, functions);

let mut module = Module::new();
module
    .section(&tables)
    .section(&elements);

let wasm_bytes = module.finish();

Implementations

impl ElementSection

pub fn new() -> Self

Create a new element section encoder.

pub fn len(&self) -> u32

The number of element segments in the section.

pub fn is_empty(&self) -> bool

Determines if the section is empty.

pub fn segment<'a>(&mut self, segment: ElementSegment<'a>) -> &mut Self

Define an element segment.

pub fn active( &mut self, table_index: Option<u32>, offset: &ConstExpr, elements: Elements<'_>, ) -> &mut Self

Define an active element segment.

Active element specifying a None table forces the MVP encoding and refers to the 0th table holding funcrefs. Non-None tables use the encoding introduced with the bulk memory proposal and can refer to tables with any valid reference type.

pub fn passive<'a>(&mut self, elements: Elements<'a>) -> &mut Self

Encode a passive element segment.

Passive segments are part of the bulk memory proposal.

pub fn declared<'a>(&mut self, elements: Elements<'a>) -> &mut Self

Encode a declared element segment.

Declared segments are part of the bulk memory proposal.

pub fn raw(&mut self, raw_bytes: &[u8]) -> &mut Self

Copy a raw, already-encoded element segment into this elements section.