Trait wasm_encoder::ComponentSection

pub trait ComponentSection: Encode {
    fn id(&self) -> u8;

    fn append_to_component(&self, dst: &mut Vec<u8>) { ... }
}

A WebAssembly component section.

Various builders defined in this crate already implement this trait, but you can also implement it yourself for your own custom section builders, or use RawSection to use a bunch of raw bytes as a section.

Required Methods

fn id(&self) -> u8

Gets the section identifier for this section.

Provided Methods

fn append_to_component(&self, dst: &mut Vec<u8>)

Appends this section to the specified destination list of bytes.