Trait wasm_encoder::Section

pub trait Section: Encode {
    fn id(&self) -> u8;
}

A WebAssembly module section.

Various builders defined in this crate already implement this trait, but you can also implement it yourself for your own custom section builders, or use RawSection to use a bunch of raw bytes as a section.

Required methods

fn id(&self) -> u8

Gets the section identifier for this section.